+++ CHANGELOG.md
@@ -4,11 +4,19 @@ All notable changes to the "vscode-bedrock-definitions" extension will be docume
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [0.0.1]
+## [1.0.0] - 2019-09-10
 
-- Initial release
+- Added autocomplete
+  - Full list of completions available in the README
+- Jump to definition for client entity identifiers, if available
+- Rewrite backend to be able to be shared by multiple providers
+  - Move to jsonc-parser from json-source-map
 
-## [0.0.2]
+## [0.0.2] - 2019-09-05
 
 - Add icon
 - Support for jsonc (JSON with comments)
+
+## [0.0.1] - 2019-09-03
+
+- Initial release